+++ src/main/java/org/hidog/walking/controllers/WalkingController.java
@@ -10,6 +10,7 @@ import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.ModelAttribute;
 import org.springframework.web.bind.annotation.RequestMapping;
 
+import java.util.ArrayList;
 import java.util.List;
 
 @Controller
@@ -28,14 +29,17 @@ public class WalkingController {
 
     @GetMapping("/map")
     public String getMainMap(Model model) {
+        List<String> addScript = new ArrayList<>();
+
         List<Double> startMarker = mainMapMarkerService.startMarkerLocation();
+
+        addScript.add("walking/mainMapMark");
+        addScript.add("walking/mainMap");
+
+        model.addAttribute("addScript", addScript);
         model.addAttribute("startMarker", startMarker);
         model.addAttribute("addCommonCss", List.of("map"));
-//        model.addAttribute("addCommonScript", List.of("map"));
-        model.addAttribute("addScript", List.of("mainMap/mainMap"));
-
 
         return utils.tpl("walking/map");
     }
 }
-//[37.560154248445315, 126.94330930709881, 37.557251187359554, 126.9434273242955, 37.55736175523448, 126.94284796714825, 37.55748082814708, 126.94234371185348]
