+++ src/main/resources/static/front/js/walking/mainMap.js
